+++ Controls-demo/treeGridNew/Expander/Expander.wml
@@ -26,13 +26,4 @@
         </Router.router:Reference>
         <Controls-demo.treeGridNew.Expander.ExpanderIcon.Index/>
     </div>
-
-    <div class="controlsDemo__cell">
-        <Router.router:Reference attr:class="controls-text-label"
-                                 state="/Controls-demo/app/:app"
-                                 app="{{ 'Controls-demo/treeGridNew/Expander/WithoutRootExpander/Index' }}">
-            <a href="{{content.href}}">Без экспандера первого уровня</a>
-        </Router.router:Reference>
-        <Controls-demo.treeGridNew.Expander.WithoutRootExpander.Index/>
-    </div>
 </div>
